+++ TP2-CC-Music Server/src/DataBase/Desafio.java
@@ -300,7 +300,7 @@ public class Desafio {
         return true;
     }
 
-    public Pergunta getPergunta(short numero) {
+    public Pergunta getPergunta(int numero) {
         return this.perguntas.get(numero - 1);
     }
 }
